+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 v2.4.2
 ------
 * Fixed detection of code flowing through the end of a method when the code at the end is dead.
+* Type information from a freshly made class is now still available after the class is finished.
 
 v2.4.1 (2022-07-24)
 ------